--- src/constants/index.js
@@ -1,2 +1,0 @@
-export * from './app'
-export * from './error'
