--- src/utils/index.js
@@ -1,2 +1,0 @@
-export * from './validator'
-export * from './number'
